Initial test example code for Firebase Authentication
Example code will run a series of tests that is then outputted to a console window. Note that the application will not work if the "google-services.json" file is not on the root folder path with the exe of the application. There is no google-services file in repo because that file is unique for each firebase project. Those who use repo must provide their own google-services file.

To get your google-services file:
1. go to Firebase
2. Console
3. Your Project
4. Project Settings
5. Your apps
6.  Android Apps (Even for Desktop projects)
7.  On "SDK setup and configuration" box, there should be a hyperlink text called google-services.json for you to click which will automatically download. Copy that file to the exe of the program after compiling.