Added error checking & visuals
Added functions to grab firebase errors for when trying to create a account. This is to allow the launcher forms to create error message boxes for when an account fails to be made, or was successful. This will later be applied to the other parts of firebase soon.

Added new background images to the launcher. The launcher randomly chooses a image to use everytime it launches.

Added audio manager to play music in the background while the launcher is running.